working on posts

+++ posts/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, reverse
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, JsonResponse
 from django.contrib.auth.decorators import login_required
+from django.core import serializers
 
 from .models import *
 from profiles.models import Profile
@@ -9,19 +10,28 @@ from .forms import PostForm
 
 
 @login_required
-def post_like(request, id):
-    post = Post.objects.get(id=id)
+def post_like(request):
+    post_id = request.POST.get('id')
+    action = request.POST.get('action')
     user = Profile.objects.get(user=request.user)
-    post.likes.create(user=user)
-    return HttpResponseRedirect(reverse('home'))
-
-
-@login_required
-def post_unlike(request, id):
-    post = Post.objects.get(id=id)
-    user = Profile.objects.get(user=request.user)
-    post.likes.get(user=user).delete()
-    return HttpResponseRedirect(reverse('home'))
+    print(post_id + action)
+    if post_id and action:
+        print('inside if')
+        try:
+            post = Post.objects.get(pk=int(post_id))
+            print(post)
+            print('got the post')
+            if action == 'like':
+                post.likes.create(user=user)
+                print('like created')
+            else:
+                post.likes.get(user=user).delete()
+                print('like deleted')
+            return JsonResponse({'status': 'ok'})
+        except:
+            print('did not got the post')
+            pass
+    return JsonResponse({'status': 'error'})
 
 
 @login_required
@@ -41,14 +51,17 @@ def post_unsave(request, id):
 
 
 @login_required
-def post_comment(request, id):
+def post_comment(request):
     profile = Profile.objects.get(user=request.user)
-    post = Post.objects.get(id=id)
-    comment = request.POST.get('text')
-    if comment:
-        post.comments.create(user=profile, text=comment)
-        return HttpResponseRedirect(reverse('home'))
-    return HttpResponseRedirect(reverse('home'))
+    if request.is_ajax and request.method == 'POST':
+        post = Post.objects.get(id=request.POST.get('post_id'))
+        comment = request.POST.get('text')
+        new_comment = post.comments.create(user=profile, text=comment)
+        new_comment.save()
+        ser_instance = serializers.serialize(
+            'json', [new_comment, profile, request.user])
+        return JsonResponse({"instance": ser_instance}, status=200)
+    return JsonResponse({"error": ""}, status=400)
 
 
 def post_delete(request):
